Start cleaning up callback implementation
== DETAILS
basic idea: instead of creating a stringified lookup and trying to find it in global(),
just pass the callback function reference itself and use that.

Note that we've only updated some of the `add_callback()` invocations so this is
broke as hell.

NEXT STEP: look for all the `add_callback` calls and make sure they're all passing functions.